Documented testing framework
