+++ server/main/test/com/smartnote/server/UploadTest.java
@@ -2,9 +2,13 @@ package com.smartnote.server;
 
 import static org.junit.Assert.*;
 
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
 import org.junit.*;
 
 import com.smartnote.server.api.v1.Upload;
+import com.smartnote.server.auth.Session;
 import com.smartnote.testing.RouteTest;
 
 import spark.Response;
@@ -21,15 +25,35 @@ public class UploadTest extends RouteTest {
     
     private Upload upload;
 
+    /**
+     * Perform the test on Upload. Tests that the response code is the expected
+     * code and that the response is in JSON format. If the response code is 200,
+     * the test also checks that the file was uploaded.
+     * 
+     * @param code the expected response code.
+     * @return the response.
+     * @throws Exception if an error occurs.
+     */
     private Response doTest(int code) throws Exception {
         Response response = handle(upload);
-        assertEquals(code, response.status());
-        assertEquals("application/json", response.type());      
+        int status = response.status();
+
+        assertEquals(code, status);
+
+        assertEquals("application/json", response.type());
         assertTrue(responseJson().has("message"));
+
+        // if response is OK, check that the file was uploaded
+        if (status == 200) {
+            Session session = getSession();
+            Path uploadPath = session.pathInSession(Paths.get("uploads", TEST_FILE_NAME));
+            assertTrue(getFileSystem().exists(uploadPath));
+        }
+
         return response;
     }
 
-    @Before
+    @Override
     public void setUp() throws Exception {
         super.setUp();
         upload = new Upload();
@@ -38,10 +62,10 @@ public class UploadTest extends RouteTest {
         setRequestQueryParam("name", TEST_FILE_NAME);
         setRequestBody(TEST_FILE_CONTENTS);
         setRequestContentType("application/pdf");
-        startRequestSession();
+        activateSession();
     }
-
-    @After
+    
+    @Override
     public void tearDown() throws Exception {
         super.tearDown();
     }
@@ -65,7 +89,7 @@ public class UploadTest extends RouteTest {
 
     @Test
     public void testUploadNoSession() throws Exception {
-        removeRequestSession();
+        deactivateSession();
         doTest(401);
     }
