+++ server/main/test/com/smartnote/testing/ExitException.java
@@ -1,7 +1,12 @@
 package com.smartnote.testing;
 
 /**
- * Used to override virtual machine exit calls.
+ * <p>Used to override virtual machine exit calls. Thrown
+ * by the <code>SecurityManager</code> when <code>System.exit()</code>
+ * is called.</p>
+ * 
+ * @author Ethan Vrhel
+ * @see BaseTest
  */
 public class ExitException extends RuntimeException {
     private final int status;
@@ -11,10 +16,15 @@ public class ExitException extends RuntimeException {
     }
 
     public ExitException(int status) {
-        super();
+        super("System.exit() was called with code " + status);
         this.status = status;
     }
 
+    /**
+     * Gets the status code passed to System.exit().
+     * 
+     * @return the status code.
+     */
     public int getStatus() {
         return status;
     }
