+++ server/main/test/com/smartnote/testing/RouteTest.java
@@ -3,52 +3,52 @@ package com.smartnote.testing;
 import static org.mockito.Matchers.*;
 import static org.mockito.Mockito.*;
 
-import java.lang.reflect.Field;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
 
 import com.google.gson.JsonObject;
-import com.smartnote.server.Server;
-import com.smartnote.server.auth.Session;
-import com.smartnote.server.auth.SessionManager;
-import com.smartnote.server.auth.SessionPermission;
-import com.smartnote.server.resource.ResourceConfig;
-import com.smartnote.server.resource.ResourceSystem;
-import com.smartnote.server.util.FileUtils;
 
 import spark.Request;
 import spark.Response;
 import spark.Route;
 
 /**
- * Used for testing Spark routes.
+ * <p>Used for testing Spark routes. This class creates mock <code>Request</code>
+ * and <code>Response</code> instances and provides methods for setting the
+ * request parameters and handling the route.</p>
+ * 
+ * <p>Not all methods are implemented. If you need to use a method that is not
+ * implemented, add it to the class and submit a pull request.</p>
  * 
  * @author Ethan Vrhel
+ * @see spark.Route
+ * @see spark.Request
+ * @see spark.Response
  */
-public class RouteTest extends BaseTest {
-    public static final String SESSION_TOKEN = "test";
-
+public class RouteTest extends BaseServerTest {
+    
+    // request
     private Request request;
     private Map<String, String> requestQueryParams;
     private String requestBody;
     private String requestContentType;
-    private Session requestSession;
 
+    // response
     private Map<String, String> responseHeaders;
     private String responseBody;
     private String responseType;
     private int responseStatus;
 
+    /**
+     * Sets up the framework for testing routes.
+     */
+    @Override
     public void setUp() throws Exception {
         super.setUp();
-
-        setupServer();
-
         this.request = mockRequest();
     }
 
+    @Override
     public void tearDown() throws Exception {
         super.tearDown();
     }
@@ -72,10 +72,6 @@ public class RouteTest extends BaseTest {
         return response;
     }
 
-    public Request request() {
-        return request;
-    }
-
     /**
      * Parse response body as JSON.
      * 
@@ -85,133 +81,99 @@ public class RouteTest extends BaseTest {
         return getGson().fromJson(responseBody, JsonObject.class);
     }
 
+    /**
+     * Sets a query parameter for the request.
+     * 
+     * @param key the key.
+     * @param value the value.
+     */
     public void setRequestQueryParam(String key, String value) {
         requestQueryParams.put(key, value);
     }
 
+    /**
+     * Removes a query parameter from the request.
+     * 
+     * @param key the key.
+     */
     public void removeRequestQueryParam(String key) {
         requestQueryParams.remove(key);
     }
 
+    /**
+     * Sets the request body.
+     * 
+     * @param requestBody the request body.
+     */
     public void setRequestBody(String requestBody) {
         this.requestBody = requestBody;
     }
 
+    /**
+     * Sets the request content type.
+     * 
+     * @param requestContentType the request content type.
+     */
     public void setRequestContentType(String requestContentType) {
         this.requestContentType = requestContentType;
     }
-
-    public void startRequestSession() {
-        this.requestSession = mockSession();
-    }
-
-    public void removeRequestSession() {
-        this.requestSession = null;
-    }
-
-    private void setupServer() throws Exception {
-        Server server = Server.getServer();
-        Class<Server> serverClass = Server.class;
-
-        Field resourceSystemField = serverClass.getDeclaredField("resourceSystem");
-        resourceSystemField.setAccessible(true);
-        resourceSystemField.set(server, createResourceSystem());
-
-        Field sessionManagerField = serverClass.getDeclaredField("sessionManager");
-        sessionManagerField.setAccessible(true);
-        sessionManagerField.set(server, mockSessionManager());
-    }
-
-    private ResourceSystem createResourceSystem() throws Exception {
-        ResourceConfig config = new ResourceConfig();
-        ResourceSystem resourceSystem = new ResourceSystem(config);
-
-        Class<ResourceSystem> resourceSystemClass = ResourceSystem.class;
-        Field fileResourceFactoryField = resourceSystemClass.getDeclaredField("fileResourceFactory");
-        fileResourceFactoryField.setAccessible(true);
-        fileResourceFactoryField.set(resourceSystem, getFileSystem().createResourceFactory());
-
-        return resourceSystem;
-    }
-
-    private SessionManager mockSessionManager() {
-        SessionManager sessionManager = mock(SessionManager.class);
-
-        when(sessionManager.getSession(any(Request.class))).thenAnswer(invokation -> requestSession);
-        when(sessionManager.createSession()).thenAnswer(invokation -> requestSession);
-        when(sessionManager.isTokenValid(anyString())).thenAnswer(invokation -> invokation.getArguments()[0].equals("test"));
-
-        return sessionManager;
-    }
     
-    private Session mockSession() {
-        Path sessionDirectory = Server.getServer().getResourceSystem().getSessionDir().resolve(SESSION_TOKEN);
-        Path tokenFile = sessionDirectory.resolve(".token");
-
-        SessionPermission permission = mock(SessionPermission.class);
-        Session session = mock(Session.class);
-
-        when(session.getId()).thenReturn(SESSION_TOKEN);
-        when(session.getSessionDirectory()).thenReturn(sessionDirectory);
-
-        doAnswer(invokation -> {
-            Path path = (Path) invokation.getArguments()[0];
-            path = sessionDirectory.resolve(path);
-            if (!FileUtils.isPathInDirectory(path, sessionDirectory))
-                throw new SecurityException("Access denied");
-            return path;
-        }).when(session).pathInSession(any(Path.class));
-
-        doAnswer(invokation -> {
-            Response response = (Response) invokation.getArguments()[0];
-            response.header("Authorization", SESSION_TOKEN);
-            return null;
-        }).when(session).writeToResponse(any(Response.class));
-
-        when(session.getPermission()).thenReturn(permission);
-        when(permission.getSession()).thenReturn(session);
-
-        return session;
-    }
-
+    // creates a mock request
     private Request mockRequest() {
         Request request = mock(Request.class);
 
+        // Request.queryParams(String)
         doAnswer(invokation -> {
             return requestQueryParams.get(invokation.getArguments()[0]);
         }).when(request).queryParams(anyString());
+
+        // Request.body()
         when(request.body()).thenAnswer(invokation -> requestBody);
+
+        // Request.bodyAsBytes()
         when(request.bodyAsBytes()).thenAnswer(invokation -> requestBody == null ? null : requestBody.getBytes());
+
+        // Request.contentType()
         when(request.contentType()).thenAnswer(invokation -> requestContentType);
         
         this.requestQueryParams = new HashMap<>();
         return request;
     }
 
+    // creates a mock response
     private Response mockResponse() {
         responseHeaders = new HashMap<>();
 
         Response response = mock(Response.class);
 
+        // Response.body(String)
         doAnswer(invokation -> {
             responseBody = (String) invokation.getArguments()[0];
             return null;
         }).when(response).body(anyString());
 
+        // Response.type(String)
         doAnswer(invokation -> {
             responseType = (String) invokation.getArguments()[0];
             return null;
         }).when(response).type(anyString());
 
+        // Response.status(int)
         doAnswer(invokation -> {
             responseStatus = (int) invokation.getArguments()[0];
             return null;
         }).when(response).status(anyInt());
 
+        // Response.body()
         when(response.body()).thenAnswer(invokation -> responseBody);
+
+        // Response.type()
         when(response.type()).thenAnswer(invokation -> responseType);
+
+        // Response.status()
         when(response.status()).thenAnswer(invokation -> responseStatus);
 
+        // Response.header(String, String)
         doAnswer(invokation -> {
             responseHeaders.put((String) invokation.getArguments()[0], (String) invokation.getArguments()[1]);
             return null;
